Glade program 2 (#1)
* Started 2nd Prototype Example

* Added more content to the second example

It is not finished yet. I am adding it to finish off later.

* Started Java Program Example

I haven't finished it. I just set up its build environment.

* Added One Line

If I'll be honest, this commit is useless.

* Finished the second example however

However it is not perfect. There are plenty of parts of the code that could be improved such as error handling.